processors/requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
textblob